move RabbitMessage
